Herramientas_Ultima_Generacion/airbnb-bigdata-project-main/src/functions/resources/requirements.txt
flask==2.0.2
google-cloud-firestore==2.3.4
google-cloud-error-reporting==1.4.1